State machine implementation

+++ source/hal/hal_dac.c
@@ -33,7 +33,7 @@ void hal_dac_set_value(hal_dac_dev_t dev, uint16_t value)
     drv->set_value(dev, value);
 }
 
-void hal_dac_set_freq(hal_dac_dev_t dev, uint16_t freq)
+void hal_dac_set_freq(hal_dac_dev_t dev, float freq)
 {
     drv->set_freq(dev, freq);
 }
